Make `Prediction` model readonly in Admin
When a model has a ForeignKey, as Prediction does onto Minute, the
Django admin tries to list every possible value in a <select> dropdown.
This doesn't work so well with 500,000 possibilities.

At the same time as fixing that, make the Prediction completely readonly
as it's safer if it's only modified through the `import_predictions`
command.

https://www.pivotaltracker.com/story/show/76505816 [Delivers #76505816]