Allow disabling particular alerts for a Location
Each location currently has alerting roughly like this:

- tide predictions: must run > 30 days into future
- surge predictions: must have every minute in the next 12 hours
- observations: must have at least 1 observation in last 60 minutes

We need to be able to turn off particular alerts for when we add a new
location. This means we have time to load tide predictions and configure
the surge & measurement collectors.

This commit allows temporarily disabling alerts, ie setting a
`tide_predictions_alerts_disabled_until` field - to make it a little bit
harder to forget to turn alerting back on!

https://www.pivotaltracker.com/story/show/80543916 [Delivers #80543916]