Add the concept of token authentication to the API
Install the `rest_framework.authtoken` app which allows you to add
authentication tokens for particular users. This will allow us to make
eg a `surge-model-collector` user with their own token.

We've currently got only read-only endpoints so this shouldn't have any
effect, but it is required for implementing the writeable
surge/measurement endpoints.

[#78037278] [#80542816] [#78441914]