(Mostly) delete old surge-model-converter app
- Migration to kill off the database columns (ModelLocation)
- Delete basically all the code.

Note that you have to keep a skeleton app with the `migrations` and
`models` directory (but without any models) in order for the
`DeleteMigration` migration to be recognised. ie if you just delete the
app, its data would hang around in the database forever.

In the future (once all databases have been migrated) we can delete the
app completely and remove it from `INSTALLED_APPS`

https://www.pivotaltracker.com/story/show/81100150 [Delivers #81100150]